# api/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: QuiverKeep Core API
  version: v1.0.0
paths:
...
  /api/v1/usage:
    get:
      summary: Usage summary
      parameters:
        - name: service
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Usage payload
...
  /api/v1/proxy/anthropic/messages:
    post:
      summary: Anthropic API proxy (L7)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Anthropic `/v1/messages` request payload passthrough.
              additionalProperties: true
      responses:
        "200":
          description: Proxied Anthropic response payload
        "400":
          description: Validation error (invalid method/content-type/body)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Core auth error (missing/invalid bearer token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: Upstream provider failure mapped to stable core error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Proxy is disabled or provider key is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "504":
          description: Upstream provider timeout mapped to stable core error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Stable error code
        message:
          type: string
          description: Human-readable error message
...